Rename `UserInfo` contract folder to `Compound`
