+++ backend/CircleForms/Contracts/ContractModels/Response/Compound/PostResponseUserInfoResponseContract.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
-using CircleForms.Contracts.ContractModels.Response.UserInfoContracts.Abstract;
+using CircleForms.Contracts.ContractModels.Response.Compound.Abstract;
 
-namespace CircleForms.Contracts.ContractModels.Response.UserInfoContracts;
+namespace CircleForms.Contracts.ContractModels.Response.Compound;
 
 public class PostUserAnswerResponseContract : UserInfoResponseContract<List<UserAnswerContract>, PostResponseContract>
 {
