+++ backend/CircleForms/Contracts/ContractModels/Response/PageResponseContract.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using CircleForms.Contracts.ContractModels.Response.UserInfoContracts.Abstract;
+using CircleForms.Contracts.ContractModels.Response.Compound.Abstract;
 
 namespace CircleForms.Contracts.ContractModels.Response;
